Fixed bug with Bus Pirate interface reset sequence.
If the Bus Pirate was placed into BitBang mode and then sent the Terminal reset command, it would brick either itself or the low level serial driver on Windows.
The solution was to re-order the commands so that it tried the BitBang reset first, and if that failed, proceeded to use the Terminal reset. Now the Bus Pirate
driver appears to reliably enter and exit BitBang mode and reset the device correctly.